Update DAO layer classes to accomodate transactions
