+++ modules/broker-common/src/main/java/org/wso2/broker/common/util/function/ThrowingFunction.java
@@ -20,23 +20,25 @@
 package org.wso2.broker.common.util.function;
 
 /**
- * Represents an operation that accepts a single input argument and returns no
- * result.
+ * Represents a function that accepts one argument and produces a result.
  *
  * <p>This is a <a href="package-summary.html">functional interface</a>
- * whose functional method is {@link #accept(Object)}.
+ * whose functional method is {@link #apply(Object)}.
  *
- * @param <T> the type of the input to the operation
- * @param <E> the type of the exception thrown
+ * @param <T> the type of the input to the function
+ * @param <R> the type of the result of the function
+ * @param <E>  the type of the exception thrown
  *
  */
 @FunctionalInterface
-public interface ThrowingConsumer<T, E extends Exception> {
+public interface ThrowingFunction<T, R, E extends Exception> {
+
     /**
-     * Performs this operation on the given argument.
+     * Applies this function to the given argument.
      *
-     * @param t the input argument.
-     * @throws E throws exception of type E on failure to consume.
+     * @param t the function argument
+     * @return the function result
+     * @throws E throws exception of type E on failure to execute function.
      */
-    void accept(T t) throws E;
+    R apply(T t) throws E;
 }
