+++ modules/broker-core/src/main/java/org/wso2/broker/core/queue/DbBackedQueueImpl.java
@@ -23,7 +23,7 @@ import org.wso2.broker.core.BrokerException;
 import org.wso2.broker.core.Message;
 import org.wso2.broker.core.Metadata;
 import org.wso2.broker.core.Queue;
-import org.wso2.broker.core.store.dao.SharedMessageStore;
+import org.wso2.broker.core.store.SharedMessageStore;
 
 import java.util.Collection;
 import java.util.concurrent.LinkedBlockingQueue;
