--- modules/broker-core/src/main/java/org/wso2/broker/core/store/dao/BaseDao.java
@@ -1,83 +1,0 @@
-/*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- *
- */
-
-package org.wso2.broker.core.store.dao;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.sql.Connection;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Statement;
-import javax.sql.DataSource;
-
-abstract class BaseDao {
-    private static final Logger LOGGER = LoggerFactory.getLogger(BaseDao.class);
-
-    private final DataSource dataSource;
-
-    BaseDao(DataSource dataSource) {
-        this.dataSource = dataSource;
-    }
-
-    public Connection getConnection() throws SQLException {
-        return dataSource.getConnection();
-    }
-
-    public void close(Connection connection) {
-        if (connection != null) {
-            try {
-                connection.close();
-            } catch (SQLException e) {
-                LOGGER.error("Error closing connection.", e);
-            }
-        }
-    }
-
-    public void close(Connection connection, Statement statement) {
-        close(statement);
-        close(connection);
-    }
-
-    public void close(Connection connection, Statement statement, ResultSet resultSet) {
-        close(resultSet);
-        close(connection, statement);
-    }
-
-    public void close(Statement statement) {
-        if (statement != null) {
-            try {
-                statement.close();
-            } catch (SQLException e) {
-                LOGGER.error("Error closing statement. " + statement, e);
-            }
-        }
-    }
-
-    public void close(ResultSet resultSet) {
-        if (resultSet != null) {
-            try {
-                resultSet.close();
-            } catch (SQLException e) {
-                LOGGER.error("Error closing result set.", e);
-            }
-        }
-    }
-}
