+++ modules/broker-core/src/main/java/org/wso2/broker/core/store/dao/impl/DaoFactory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -11,19 +11,19 @@
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
+ * KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations
  * under the License.
  *
  */
 
-package org.wso2.broker.core.store.dao;
+package org.wso2.broker.core.store.dao.impl;
 
 import org.wso2.broker.core.metrics.BrokerMetricManager;
-import org.wso2.broker.core.store.dao.impl.BindingDaoImpl;
-import org.wso2.broker.core.store.dao.impl.ExchangeDaoImpl;
-import org.wso2.broker.core.store.dao.impl.MessageDaoImpl;
-import org.wso2.broker.core.store.dao.impl.QueueDaoImpl;
+import org.wso2.broker.core.store.dao.BindingDao;
+import org.wso2.broker.core.store.dao.ExchangeDao;
+import org.wso2.broker.core.store.dao.MessageDao;
+import org.wso2.broker.core.store.dao.QueueDao;
 
 import javax.sql.DataSource;
 
@@ -45,7 +45,7 @@ public class DaoFactory {
     }
 
     public MessageDao createMessageDao() {
-        return new MessageDaoImpl(dataSource, metricManager);
+        return new MessageDaoImpl(new MessageCrudOperationsDao(dataSource, metricManager));
     }
 
     public ExchangeDao createExchangeDao() {
