+++ modules/broker-core/src/main/java/org/wso2/broker/core/store/dao/impl/NoOpBindingDao.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
  *
  * WSO2 Inc. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -11,25 +11,23 @@
  * Unless required by applicable law or agreed to in writing,
  * software distributed under the License is distributed on an
  * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
+ * KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations
  * under the License.
  *
  */
 
-package org.wso2.broker.core.store.dao;
+package org.wso2.broker.core.store.dao.impl;
 
 import org.wso2.broker.core.Binding;
+import org.wso2.broker.core.store.dao.BindingDao;
 
 /**
  * {@link BindingDao} implementation which represents the NullObject pattern. Used for non durable exchanges where
  * bindings shouldn't be persisted.
  */
-public class NoOpBindingDao extends BindingDao {
+public class NoOpBindingDao implements BindingDao {
 
-    public NoOpBindingDao() {
-        super(null);
-    }
 
     @Override
     public void persist(String exchangeName, Binding binding) {
